evalengine: virtual machine (#12369)
* go: add a slices2 package


* evalengine: compiler [wip]


* evalengine: fix test helpers


* evalengine/compiler: Add support for REPEAT


* evalengine/compiler: Add support for base64 functions


* evalengine/compiler: Implement UPPER / LOWER


* evalengine/compiler: Implement string length functions


* evalengine/compiler: Implement ASCII() function


* evalengine/compiler: Implement LIKE support


* evalengine/compiler: fix naming for instructions


* evalengine/compiler: use new collation API


* evalengine/compiler: implement multi-comparisons


* evalengine/compiler: implement JSON_ARRAY


* Fix issues after rebasing

This fixes the compiler for the newly exposed issues after rebasing on
top of evalengine fixes.


* evalengine/compiler: Implement hex() and convert using


* evalengine/compiler: implement JSON_OBJECT


* compiler: add support for tuple comparisons


* evalengine/compiler: arrange in separate files


* evalengine/compiler: Implement COLLATION function


* evalengine/compiler: Implement CEIL


* evalengine/compiler: IS expressions and JSON_KEYS


* evalengine: fix IS and CEIL corner cases


* evalengine/compiler: implement IN


* evalengine/compiler: implement WEIGHT_STRING


* evalengine/compiler: optimize null checks


* evalengine/compiler: test for runtime errors


* evalengine: Add support for FLOOR().

We already had CEIL() and this is very similar, in fact, it's a tiny bit
simpler since we don't have to do the add one logic if divmod returns a
non zero remainder.


* evalengine: Add support for ABS()

Since we already have CEIL(), FLOOR() etc. let's add some more numeric
operations here to the eval engine.


* evalengine: Implement trigonometry functions


* evalengine: simplify test case generation


* evalengine: add missing headers


* evalengine: sizegen


* slices2: update licensing


* evalengine: add documentation


* Move normalization logic as it's test only


* evalengine/compiler: Add missing COT to compiler.


* evalengine/compiler: Fix flags for arithmatic operations


* refactor comparisons


* no need to pass in err to method


* evalengine/compiler: remove redundant constants


* evalengine/compiler: de-duplicate math functions


* evalengine: Add additional arithmetic tests and fix edge cases

This also fixes how we deal with boolean values so we don't rewrite
internal booleans on the stack and update what literal booleans mean
accidentally.


---------

+++ go/vt/vtgate/evalengine/api_arithmetic_test.go
@@ -350,7 +350,7 @@ func TestArithmetics(t *testing.T) {
 			// testing for overflow of float64
 			v1:  NewFloat64(math.MaxFloat64),
 			v2:  NewFloat64(0.5),
-			err: dataOutOfRangeError(math.MaxFloat64, 0.5, "BIGINT", "/").Error(),
+			err: dataOutOfRangeError(math.MaxFloat64, 0.5, "DOUBLE", "/").Error(),
 		}},
 	}, {
 		operator: "*",
